Store grading latency of each submission
This adds a new column scored_at to the submission_results table,
which stores the (first) time the submission is scored.

Then, the grading  latency is defined as the duration between the
submission timestamp and this scored_at value. Basically, this is
how long the user waits until they see their submission verdict.