Add config to allow contestants to do unofficial submissions after contest ends
This adds a new config to a contest that allows contestants to submit solutions,
which will be unofficial submissions, between the end of the contest and
the start of the analysis mode. This is used to allow contestants to continue
working if they think they deserve extra time (which should then be accompanied
by a formal appeal).